+++ src/main/java/com/boardgo/domain/mapper/MeetingMapper.java
@@ -9,6 +9,7 @@ import com.boardgo.domain.meeting.entity.enums.MeetingType;
 import com.boardgo.domain.meeting.repository.projection.HomeMeetingDeadlineProjection;
 import com.boardgo.domain.meeting.repository.projection.LikedMeetingMyPageProjection;
 import com.boardgo.domain.meeting.repository.projection.MeetingDetailProjection;
+import com.boardgo.domain.meeting.repository.projection.MeetingReviewProjection;
 import com.boardgo.domain.meeting.repository.projection.MeetingSearchProjection;
 import com.boardgo.domain.meeting.repository.projection.MyPageMeetingProjection;
 import com.boardgo.domain.meeting.service.response.HomeMeetingDeadlineResponse;
@@ -20,6 +21,7 @@ import com.boardgo.domain.meeting.service.response.MeetingSearchPageResponse;
 import com.boardgo.domain.meeting.service.response.MeetingSearchResponse;
 import com.boardgo.domain.meeting.service.response.MyPageMeetingResponse;
 import com.boardgo.domain.meeting.service.response.UserParticipantResponse;
+import com.boardgo.domain.review.service.response.ReviewMeetingResponse;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -162,4 +164,7 @@ public interface MeetingMapper {
 
     List<HomeMeetingDeadlineResponse> toHomeMeetingDeadlineResponses(
             List<HomeMeetingDeadlineProjection> homeMeetingDeadlineProjections);
+
+    List<ReviewMeetingResponse> toReviewMeetingResponses(
+            List<MeetingReviewProjection> meetingReviewProjection);
 }
