+++ src/main/java/com/boardgo/domain/meeting/service/MeetingQueryUseCase.java
@@ -8,6 +8,7 @@ import com.boardgo.domain.meeting.service.response.LikedMeetingMyPageResponse;
 import com.boardgo.domain.meeting.service.response.MeetingDetailResponse;
 import com.boardgo.domain.meeting.service.response.MeetingSearchResponse;
 import com.boardgo.domain.meeting.service.response.MyPageMeetingResponse;
+import com.boardgo.domain.review.service.response.ReviewMeetingResponse;
 import java.util.List;
 import java.util.Map;
 
@@ -37,4 +38,9 @@ public interface MeetingQueryUseCase {
     List<MeetingEntity> findByIdIn(List<Long> meetingIdList);
 
     List<LikedMeetingMyPageResponse> findLikedMeeting(List<Long> meetingIdList);
+
+    List<MeetingEntity> findAllById(List<Long> meetingIds);
+
+    List<ReviewMeetingResponse> findReviewableMeeting(
+            Long reviewerId, List<Long> finishedReviewMeetingIds);
 }
