+++ src/main/java/com/boardgo/domain/review/service/ReviewQueryUseCase.java
@@ -1,15 +1,15 @@
 package com.boardgo.domain.review.service;
 
-import com.boardgo.domain.review.entity.enums.ReviewType;
 import com.boardgo.domain.review.service.response.MyEvaluationTagsResponse;
 import com.boardgo.domain.review.service.response.MyReviewsResponse;
-import com.boardgo.domain.review.service.response.ReviewMeetingResponse;
 import com.boardgo.domain.review.service.response.ReviewMeetingReviewsResponse;
 import java.util.List;
 
 public interface ReviewQueryUseCase {
 
-    List<ReviewMeetingResponse> getReviewMeetings(ReviewType reviewType, Long userId);
+    List<Long> findMeetingIdsOfWrittenReview(Long userId);
+
+    List<Long> findFinishedReviewMeetingIds(Long userId);
 
     List<Long> getReviewMeetingParticipants(Long meetingId, Long reviewerId);
